+++ packages/apps/dev-wallet/src/App/Layout/style.css.ts
@@ -28,3 +28,12 @@ export const mainContainerClass = style({
     },
   },
 });
+
+export const mobileNetworkClass = style({
+  display: 'flex',
+  '@media': {
+    'screen and (min-width: 640px)': {
+      display: 'none',
+    },
+  },
+});