+++ packages/apps/dev-wallet/src/pages/transaction/components/TxPipeLine.tsx
@@ -221,15 +221,29 @@ function TxStatusList({
               {sendDisabled ? 'Transaction is pending' : 'Ready to preflight'}
             </Stack>
           </Text>
+          <Text size="small">
+            Preflight will test your transaction first to avoid paying gas for a
+            failed submission.
+          </Text>
           {variant === 'expanded' && (
-            <Button
-              isCompact
-              onClick={() => onPreflight()}
-              isDisabled={sendDisabled}
-              startVisual={<MonoViewInAr />}
-            >
-              Preflight transaction
-            </Button>
+            <Stack gap={'sm'}>
+              <Button
+                isCompact
+                onClick={() => onPreflight()}
+                isDisabled={sendDisabled}
+                startVisual={<MonoViewInAr />}
+              >
+                Preflight
+              </Button>
+              <Button
+                variant="outlined"
+                startVisual={<MonoShare />}
+                isCompact
+                onClick={copyTx}
+              >
+                {copied ? 'copied' : 'Share'}
+              </Button>
+            </Stack>
           )}
         </Stack>
       ),
@@ -282,14 +296,24 @@ function TxStatusList({
             </Stack>
           </Text>
           {variant === 'expanded' && (
-            <Button
-              isCompact
-              onClick={() => onSubmit()}
-              isDisabled={sendDisabled}
-              startVisual={<MonoViewInAr />}
-            >
-              Send transaction
-            </Button>
+            <Stack gap={'sm'}>
+              <Button
+                isCompact
+                onClick={() => onSubmit()}
+                isDisabled={sendDisabled}
+                startVisual={<MonoViewInAr />}
+              >
+                Send tx
+              </Button>
+              <Button
+                variant="outlined"
+                startVisual={<MonoShare />}
+                isCompact
+                onClick={copyTx}
+              >
+                {copied ? 'copied' : 'Share'}
+              </Button>
+            </Stack>
           )}
         </Stack>
       ),