+++ packages/apps/dev-wallet/src/pages/transfer/Steps/TransferForm.tsx
@@ -406,14 +406,32 @@ export function TransferForm({
     onSubmit(
       {
         ...data,
+        receivers: [
+          ...data.receivers,
+          ...xchainSameAccount.map((x) => ({
+            amount: x.amount,
+            address: senderAccount.address,
+            chain: x.target,
+            xchain: true,
+            chunks: [
+              {
+                amount: x.amount,
+                chainId: x.source,
+              },
+            ],
+            discoveredAccount: senderAccount,
+          })),
+        ],
         gasPayer: data.gasPayer || data.senderAccount,
         creationTime: data.creationTime ?? Math.floor(Date.now() / 1000),
       },
-      [...xchainSameAccount, ...redistribution],
+      [...redistribution],
     );
   }
 
   const senderChain = watch('chain');
+
+  const isSafeTransfer = watch('type') === 'safeTransfer';
 
   return (
     <form onSubmit={handleSubmit(onSubmitForm)}>
@@ -648,9 +666,7 @@ export function TransferForm({
                                     <Stack flexDirection={'column'}>
                                       <AccountSearchBox
                                         accounts={filteredAccounts}
-                                        hideKeySelector={
-                                          getValues('type') !== 'safeTransfer'
-                                        }
+                                        hideKeySelector={!isSafeTransfer}
                                         watchedAccounts={
                                           filteredWatchedAccounts
                                         }