+++ packages/apps/rwa-demo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @kadena/rwa-demo
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies \[35cad4d]
+- Updated dependencies \[ebc4e5c]
+  - @kadena/graph\@2.1.0
+  - @kadena/kode-ui\@0.20.1
+
 ## 0.1.6
 
 ### Patch Changes